% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_ag.R
\docType{data}
\name{base_ag}
\alias{base_ag}
\title{Base de ejemplo para psaLAC}
\format{
Un data frame con 12 filas y 3 columnas:
\describe{
\item{anio}{integer: año (2020–2022)}
\item{sexo}{character: 'H' o 'M'}
\item{poblacion}{integer: población}
}
}
\usage{
data(base_ag)
}
\description{
Conjunto de datos mínimo de población por año y sexo para ejemplos y pruebas.
}
\examples{
data(base_ag)
head(base_ag)
}
\keyword{datasets}
