% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa_context.R
\name{psa_context}
\alias{psa_context}
\title{Título.}
\usage{
psa_context(
  data,
  var_pop,
  var_terr = NULL,
  var_year,
  var_sex,
  var_age,
  var_area,
  year_piramid = NULL
)
}
\arguments{
\item{data}{p1.}

\item{var_pop}{p2.}

\item{var_terr}{p3.}

\item{var_year}{p4.}

\item{var_sex}{p5.}

\item{var_age}{p6.}

\item{var_area}{p7.}

\item{year_piramid}{p8.}
}
\value{
ext_pol
}
\description{
Descripción.
}
\details{
Detalles. Detalles.
}
\examples{
psa_context(data = base_ag, var_pop = "poblacion", var_terr = NULL, var_year = "anio", var_sex = "sexo", var_age = "edad", var_area = "area", year_piramid = NULL)
}
\references{
Gutierrez, H. A. (2009), \emph{Estrategias de muestreo: Diseno de encuestas y estimacion de parametros}. Editorial Universidad Santo Tomas.
Valliant, R, et. al. (2013), \emph{Practical tools for Design and Weighting Survey Samples}. Springer
}
\author{
Angel Gaibor <mat.angel.gaibor at gmail.com>

Javier Núñez <mat.javier.nunez at gmail.com>
}
