% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa_context.R
\name{psa_context}
\alias{psa_context}
\title{Función contexto de psaLAC}
\usage{
psa_context(data, var_pop, var_terr = NULL, var_year, var_sex)
}
\arguments{
\item{data}{una base de datos de población}

\item{var_pop}{una variable que denota var_pop}

\item{var_terr}{una variable que denota var_terr}

\item{var_year}{una variable que denota var_year}

\item{var_sex}{una variable que denota var_sex}
}
\value{
una lista con los df calculados a partir de \code{data}, \code{var_pop}, \code{var_terr}, \code{var_year} y \code{var_sex}.
}
\description{
Esta funcion calculo el contexto pais usando psaLAC
}
\examples{
psa_context(base_ag, var_pop = "poblacion", var_year = "anio", var_sex = "sexo")
}
