% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{package}
\name{psaLAC-package}
\alias{psaLAC}
\alias{psaLAC-package}
\title{psaLAC: Función prueba psaLAC}
\description{
Un paquete para calcular el análisis de situación poblacional. Mi mamá me mima. Mi papá me ama.
}
\author{
\strong{Maintainer}: Javier Nuñez \email{mat.javier.nunez@gmail.com}

}
\keyword{internal}
